+++ templates/index.php
@@ -1,6 +1,6 @@
 <?php
 $appId = OCA\ElectronicSignatures\AppInfo\Application::APP_ID;
-OCP\Util::addscript($appId, 'electronic-signatures-main');
+OCP\Util::addInitScript($appId, 'electronic-signatures-main');
 OCP\Util::addStyle($appId, 'icons');
 ?>