+++ templates/settings/admin.php
@@ -1,6 +1,6 @@
 <?php
 $appId = OCA\ElectronicSignatures\AppInfo\Application::APP_ID;
-OCP\Util::addscript($appId, 'electronic-signatures-adminSettings');
+OCP\Util::addInitScript($appId, 'electronic-signatures-adminSettings');
 /** @var array $_ */
 /** @var OCP\IL10N $l */
 ?>
@@ -15,6 +15,7 @@ OCP\Util::addscript($appId, 'electronic-signatures-adminSettings');
   data-container-type="<?php echo $_['container_type'] ?>"
   data-api-language="<?php echo $_['api_language'] ?>"
   data-remote-signing-queue-status-webhook="<?php echo $_['remote_signing_queue_status_webhook'] ?>"
+  data-default-remote-signing-queue-status-webhook="<?php echo $_['default_remote_signing_queue_status_webhook'] ?>"
   data-signing-mode="<?php echo $_['signing_mode'] ?>"
 ></div>